# src/docs/swagger/collections.yaml
openapi: 3.0.0
info:
  title: Shopease
  description: The API collection for Shopease
  version: 1.0.0
servers:
  - url: http://localhost:9000/api/v1
    description: Shopease Development Server
  - url: http://api.shopease.shubhampurwar.in/api/v1
    description: Shopease Production Server
...
paths:
...
  /orders:
    get:
      tags:
        - Orders
      description: Allows users to fetch their orders
      operationId: getOrders
      parameters:
        - $ref: '#/components/parameters/DaysInPast'
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: A JSON object containing a list of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
          headers:
            X-Total-Count:
              description: Total number of orders that belong to a user
              schema:
                type: integer

    post:
      tags:
        - Orders
      description: Allows users to place an order
      operationId: createOrder
      requestBody:
        description: A JSON object containing order info
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderBody'
      responses:
        '201':
          description: A JSON object containing ID of a newly created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ID'
  /orders/{orderId}:
    get:
      tags:
        - Orders
      description: Allows users to fetch one of their orders by ID
      operationId: getOrderById
      parameters:
        - in: path
          name: orderId
          required: true
          description: ID of the order to be fetched
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing an order
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/OrderWithAddress'
  /orders/{orderId}/confirm:
    put:
      tags:
        - Orders
      description: Confirms an order upon payment success
      operationId: confirmOrder
      parameters:
        - in: path
          name: orderId
          required: true
          description: ID of the order that is to be confirmed upon payment success
          schema:
            type: string
      requestBody:
        description: A JSON object containing razorpay payment info
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                razorpayPaymentId:
                  type: string
                razorpaySignature:
                  type: string
      responses:
        '200':
          description: A JSON object containing a confirmed order
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/OrderWithAddress'
  /orders/{orderId}/cancel:
    put:
      tags:
        - Orders
      description: Allows users to cancel their order
      operationId: cancelOrder
      parameters:
        - in: path
          name: orderId
          required: true
          description: ID of the order to be cancelled
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing a cancelled order
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Order'
  /admin/orders:
    get:
      tags:
        - Orders
      description: Allows admins to fetch a list of orders
      operationId: adminGetOrders
      parameters:
        - $ref: '#/components/parameters/DaysInPast'
        - in: query
          name: status
          description: Field to filter orders by their status
          schema:
            type: array
            items:
              type: string
              enum:
                - created
                - processing
                - shipped
                - delivered
                - cancelled
          style: form
          explode: false
        - in: query
          name: paid
          description: Field on filter orders based on whether they are paid or not
          schema:
            oneOf:
              - type: boolean
              - type: string
            enum:
              - true
              - false
              - all
          style: form
          explode: false
        - in: query
          name: sort
          description: Field to sort orders by date or amount in asc|desc order
          schema:
            type: string
            enum:
              - date_asc
              - date_desc
              - amount_asc
              - amount_desc
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: A JSON object containing a list of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderWithAddress'
          headers:
            X-Total-Count:
              description: Total number of orders
              schema:
                type: integer
  /admin/orders/{orderId}:
    get:
      tags:
        - Orders
      description: Allows admins to fetch an order by ID
      operationId: adminGetOrderById
      parameters:
        - in: path
          name: orderId
          required: true
          description: ID of the order to be fetched
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing an order
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/OrderWithAddress'

    put:
      tags:
        - Orders
      description: Allows admins to update the status of an order
      operationId: updateOrderStatus
      parameters:
        - in: path
          name: orderId
          required: true
          description: ID of the order whose status needs to be updated
          schema:
            type: string
      requestBody:
        description: A JSON object containing a new order status
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - processing
                    - shipped
                    - delivered
      responses:
        '200':
          description: A JSON object containing an order with updated status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/OrderWithAddress'

    delete:
      tags:
        - Orders
      description: Allows admins to delete an order
      operationId: deleteOrder
      parameters:
        - in: path
          name: orderId
          required: true
          description: ID of the order to be deleted
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing ID of the deleted order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ID'
...
components:
  parameters:
    Page:
      in: query
      name: page
      description: Number of records to skip
      schema:
        type: integer
        format: int32
        minimum: 1
...
    DaysInPast:
      in: query
      name: daysInPast
      description: Number of days to fetch past orders
      schema:
        type: integer
        format: int32
        min: 1
  schemas:
...
    ID:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            id:
              type: string
...
    Address:
      type: object
      properties:
        id:
          type: string
        fullname:
          type: string
        phone:
          type: string
        line1:
          type: string
        line2:
          type: string
        landmark:
          type: string
        country:
          type: string
        state:
          type: string
        city:
          type: string
        postalCode:
          type: string
        isDefault:
          type: boolean
        user:
          type: string
        isDeleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Product:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        brand:
          type: string
        category:
          type: string
        stock:
          type: integer
        image:
          type: object
          properties:
            url:
              type: string
            publicId:
              type: string
        soldUnits:
          type: integer
        avgRating:
          type: number
        reviewCount:
          type: integer
        createdBy:
          type: string
        lastUpdatedBy:
          type: string
        deletedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
        isDeleted:
          type: boolean
...
    Order:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              product:
                $ref: '#/components/schemas/Product'
              quantity:
                type: integer
              price:
                type: number
        orderAmount:
          type: number
        discount:
          type: number
        shippingCharges:
          type: number
        taxAmount:
          type: number
        totalAmount:
          type: number
        couponCode:
          type: string
        shippingAddress:
          type: string
        deliveryMode:
          type: string
          enum:
            - standard
            - express
        isPaid:
          type: boolean
        paymentId:
          type: string
        status:
          type: string
          enum:
            - created
            - processing
            - shipped
            - delivered
            - cancelled
        estimatedDeliveryDate:
          type: string
          format: date-time
        user:
          type: string
        isDeleted:
          type: boolean
        deletedBy:
          type: string
        statusLastUpdatedBy:
          type: string
        deletedAt:
          type: string
          format: date-time
        statusUpdatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderWithAddress:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              product:
                $ref: '#/components/schemas/Product'
              quantity:
                type: integer
              price:
                type: number
        orderAmount:
          type: number
        discount:
          type: number
        shippingCharges:
          type: number
        taxAmount:
          type: number
        totalAmount:
          type: number
        couponCode:
          type: string
        shippingAddress:
          $ref: '#/components/schemas/Address'
        deliveryMode:
          type: string
          enum:
            - standard
            - express
        isPaid:
          type: boolean
        paymentId:
          type: string
        status:
          type: string
          enum:
            - created
            - processing
            - shipped
            - delivered
            - cancelled
        estimatedDeliveryDate:
          type: string
          format: date-time
        user:
          type: string
        isDeleted:
          type: boolean
        deletedBy:
          type: string
        statusLastUpdatedBy:
          type: string
        deletedAt:
          type: string
          format: date-time
        statusUpdatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderBody:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              product:
                type: string
              quantity:
                type: integer
        couponCode:
          type: string
        shippingAddress:
          type: string
        deliveryMode:
          type: string
          enum:
            - standard
            - express
...